Move config logic into config_ini.rs
